+++ modules/photoeffects/src/boostColor.cpp
@@ -19,7 +19,7 @@ void boostColor(cv::InputArray src, cv::OutputArray dst, float intensity)
     Mat srcHls;
     cvtColor(srcImg, srcHls, COLOR_BGR2HLS);
 
-    int intensityInt = intensity * MAX_INTENSITY;
+    int intensityInt = (int)(intensity * MAX_INTENSITY);
     srcHls += Scalar(0, 0, intensityInt);
 
     cvtColor(srcHls, dst, COLOR_HLS2BGR);
